Fixed player card colour issue, and reselecting 'card_slot' card
After the first 'round', after confirming a card selection, the player card was not being highlighted in blue. This commit fixes the problem.

After moving on to another round after choosing a winning card, the winning card was still currently selected and could then be comfirmed - causing the game to break.